% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sframe_methods.R
\name{format.sframe}
\alias{format.sframe}
\title{Format an sframe instrument object as a string}
\usage{
\method{format}{sframe}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sframe}.}

\item{...}{Ignored. Present for S3 consistency.}
}
\value{
A single character string.
}
\description{
Format an sframe instrument object as a string
}
