% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{model_report_template}
\alias{model_report_template}
\title{Create a model reporting template}
\usage{
model_report_template(model, include_json = TRUE)
}
\arguments{
\item{model}{An \code{\link[=sf_model]{sf_model()}} object.}

\item{include_json}{Logical. Whether to include the JSON schema block.}
}
\value{
A character string.
}
\description{
Create a model reporting template
}
