% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_layer.R
\name{sf_indirect}
\alias{sf_indirect}
\title{Define an indirect effect path}
\usage{
sf_indirect(from, through, to, label = NULL)
}
\arguments{
\item{from}{Source construct ID.}

\item{through}{Character vector of mediator construct IDs.}

\item{to}{Target construct ID.}

\item{label}{Optional effect label.}
}
\value{
An object of class \code{sf_indirect}.
}
\description{
Define an indirect effect path
}
